chore: Set minimum date and time for reservation input field
This commit sets the minimum date and time for the "FechayhoradelaReserva" input field in "generaReserva.js" to the current date and time. This ensures that users cannot select a date and time in the past when making a reservation.

Note: This commit message follows the established convention of using the "chore" type for code maintenance and updates.