refactor: Update dbOperations.js with new functions
Update dbOperations.js to include the new functions seleccionarDeTabla and actualizarEnTabla. These functions allow for selecting data from a table and updating data in a table respectively. This improves code organization and reusability.

Note: This commit message follows the established convention of using the "refactor" type for code improvements and updates.