chore: Update server.js to include new functions for database operations
Update server.js to include the new functions seleccionarDeTabla and actualizarEnTabla from dbOperations.js. This allows for selecting data from a table and updating data in a table respectively. The addition of these functions improves code organization and reusability.

Note: This commit message follows the established convention of using the "chore" type for code maintenance and updates.