fix: Add min attribute to personas input field in restaurant.html
The personas input field in restaurant.html was missing the min attribute, which caused users to be able to enter negative values or zero. This commit adds the min attribute with a value of 1 to ensure that users can only enter positive values for the number of personas.

Note: This commit message follows the established convention of using the "fix" type for bug fixes.