refactor: Update perfil.html and buscarPlato.js to dynamically display restaurant dishes
This commit refactors the code in perfil.html and buscarPlato.js to enable the dynamic display of restaurant dishes on the page. The buscarPlato.js file has been added, which fetches the list of dishes for a restaurant using the restaurant's email stored in the session storage. The fetched data is then displayed in the perfil.html file, enhancing the user experience by providing up-to-date and relevant information about the restaurant's menu.

Note: This commit message follows the established convention of using the "refactor" type for code improvements and updates.