refactor: Update GeneraPerfil.js and perfil.html to remove unnecessary code and improve performance
This commit removes the unnecessary code in GeneraPerfil.js and perfil.html, specifically the line that reloads the page after setting the restaurant email in the sessionStorage. This line is redundant and can be safely removed without affecting the functionality of the code. Removing this line improves the performance of the page by eliminating an unnecessary reload.

Note: This commit message follows the established convention of using the "refactor" type for code improvements and updates.