refactor: Update ListaReservas.js to improve search functionality
This commit updates the ListaReservas.js file to improve the search functionality for reservations. The code change adds a conditional check to filter the reservations based on the search input. If there is no search input, all reservations are displayed. This enhancement improves the usability of the page by allowing users to easily find specific reservations.

Note: This commit message follows the established convention of using the "refactor" type for code improvements and updates.