refactor: Improve search functionality in ListaReservas.js
This commit refactors the ListaReservas.js file to enhance the search functionality for reservations. It adds a conditional check to filter reservations based on the search input. If there is no search input, all reservations are displayed. This improvement improves the usability of the page by allowing users to easily find specific reservations.