refactor: Disable inline styles in .hintrc
This commit disables the "no-inline-styles" hint in the .hintrc file. By setting it to "off", we allow the use of inline styles in the code. This change is made to accommodate specific styling requirements and ensure consistency throughout the project.

Note: This commit message follows the established convention of using the "refactor" type for code improvements and updates.