refactor: Add "Ver detalles" button functionality in ListaReservas.js
This commit adds functionality to the "Ver detalles" button in the ListaReservas.js file. When the button is clicked, it displays the details of the corresponding reservation, including the ID, client email, date, time, number of people, and status. This enhancement improves the user experience by providing more information about each reservation.

Note: This commit message follows the established convention of using the "refactor" type for code improvements and updates.